Rename all transaction/airdrop sender factories with the ‘Factory’ suffix
